pyproject.toml
[project]
name = "fraud-detection-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.0", # Data manuipulation and analysis 
    "pymysql>=1.1.2", # SQL driver
    "pyyaml>=6.0.3", # Read .yaml files
    "sqlalchemy>=2.0.46", # ORM
]
